Added CSRF protection to AJAX post requests.  See csrf_setup.js for added functionality.
