+++ pychart/pychart/templates/pychart/base.html
@@ -59,6 +59,7 @@
     <script src="{% static 'assets/bootstrap.min.js' %}"></script>
     <script src="//cdnjs.cloudflare.com/ajax/libs/bootstrap-table/1.11.0/bootstrap-table.min.js"></script>
     <script src="//cdnjs.cloudflare.com/ajax/libs/bootstrap-table/1.11.0/locale/bootstrap-table-en-US.min.js"></script>
+    <script src="{% static 'assets/csrf_setup.js' %}"></script>
     <script src="{% static 'assets/update_table.js' %}"></script>
     <script src="{% static 'assets/render_data.js' %}"></script>
     <script src="{% static 'assets/save_render.js' %}"></script>
